[REF] l10n_do_accounting: move environment() function to common file
